Handle PD Output deletion permissions
